move BuzzPlace to Place.Buzz
