+++ App/src/test/java/cc/softwarefactory/lokki/android/espresso/BuzzTest.java
@@ -14,7 +14,6 @@ import java.util.List;
 import cc.softwarefactory.lokki.android.R;
 import cc.softwarefactory.lokki.android.espresso.utilities.MockJsonUtils;
 import cc.softwarefactory.lokki.android.espresso.utilities.TestUtils;
-import cc.softwarefactory.lokki.android.models.BuzzPlace;
 import cc.softwarefactory.lokki.android.models.Place;
 
 import static android.support.test.espresso.Espresso.onView;
